ASCI: Forward UPLINK RELEASE on dedicated channel to MSC
If the phone is (still) on a dedicated channel, it may release the
uplink in case of a voice group call. It depends on the MSC how to
handle the situation. Currently it releases the call.

Generally the phone is assigned to the VGCS/VBS channel before it
releases the uplink.

Change-Id: Ib91c282ed36e82b38c0e738533e3a421de81a9a8
Related: OS#4852