RBS: Fix acceptance of SI2quater in System Information
The DUG20 does not seem to accept SI2quater unless we have
RSL_IE_FULL_BCCH_INFO at the end of the message.

We were already doing this for SI13, but it seems the DUG
needs it for the SI2quater also. In fact, adding this IE
to the end of ALL BCCH SI does not generate any error.

Confirmed bt capturing BCCH off the air and decoding.
All SI as sent via RSL are represented correctly on C0.

Change-Id: Icfd55a8272a42a592b1c546ab3eba8757868b6ca