osmo_msc_data_alloc(), doc/examples: deny emergency calls by default
It's safer to deny emergency calls by default, and thus require
osmo-bsc users to allow them manually iff such calls can actually
be routed to their local emergency services.

This reduces the risk of a situation, when a person having a real-life
emergency would be trying, for instance, to call an ambulance via a
test network, which does allow emergency calls but cannot route them.

Change-Id: I5296b3031c14a896caa33b16bf04db762b9f30bf