Convert bts->depends_on from bitmask to llist
The amount of dependencies on each BTS is usually zero or a small
number, hence the computation cost is mostly similar regardless of using
a llist or a bitmask buffer.

Right now, even if the dependency chains are (almost) empty, the
previous code allocated (N * 256/8) bytes, where N is the amount of bts
configured.

Since we are now planning to increase the amount of bts (N) from uint8_t to
uint16_t, that means with this change we will then avoid:
* Increasing memory use to O(N * 65536/8).
* Having to adapt code operating on bitmask (size)

Related: SYS#7062
Change-Id: I59e37d6baa020aeafdffc2c3538e988effd37620