Introduce typedef for bts->nr gsm_bts_nr_t
The size of bts->nr is implicitly or explicitly used or taken into
account in several places in the code.
Add a new type to make it easier to define its size and range in one
place.

This also helps in identifying fields contains an BSC-implementation
"bts_nr" vs theip.access Abis OML "bts_nr" sent over the wire, which is
always a uint8_t according to spec.

Related: SYS#7062
Change-Id: Ic0db2873fa9795024aba4399da495a0d8f7504b0