Log unsupported PCUIF for bts > 255
Since the "struct gsm_pcu_if"->bts_nr is still uint8_t, that means
only the first 256 configured BTS can properly make use of the
PCUIF interface and hence have gprs support.

This is not a major problem as long as the BTS connected to osmo-bsc
are of type osmo-bts, since usually those use a BTS co-located
osmo-pcu, hence on each one bts_nr=0.

But if we wanted to run them with osmo-bsc PCU co-located,
(eg rbs2000), then we'd need to extend the PCUIF interface.

Related: OS#6565
Related: SYS#7062
Change-Id: If5a8cd94195422989de3daa04be7ffc858c0a8eb