Fix missing LAC/CI hashtable update upon CTRL SET
Add functions to set LAC/CI on a BTS and update hashtables

Use these function from both the vty and the control interface.

This commit fixes the issue where an update to LAC or CI
via the control interface was not resulting in updated
hashtables and therefore the BTS was not found during paging.

Related: OS#6840
Fixes: 15284337ece94e3ff3ebf9575096290f67d69a88

Change-Id: If9509534bd4a75f8ed8fb3e0a6bec701c7596861