experiment: send extra RSL IEs for AMR per vty option
This patch is intended as an experiment, not for merging!

3GPP TS 48.058 (Abis RSL) defines two IEs, namely MultiRate Control
(not to be confused with MultiRate Configuration) and Supported
Codec Types, which need to be added to CHAN ACTIV command message
if AMR is used *and* it is desired to have working TFO-AMR per
TS 28.062.  Standard OsmoBSC never sends either IE: neither of these
IEs is implemented in OsmoBTS or in ip.access nanoBTS, and no one
prior to us has attempted to bring up AMR on classic E1 BTS.

Add vty options that enable sending of each of these IEs individually,
to facilitate experiments with different combinations on real E1 BTS
hardware.

Change-Id: Ib66bdf5f15717eb3f74fee34e0f436fac2241356