experiment: send extra RSL IEs for AMR per vty option
This patch is intended as an experiment, not for merging!

With the introduction of TFO for AMR in 3GPP Rel4, further extended
in Rel5, TS 48.058 (Abis RSL) got some new IEs which need to be
added to CHAN ACTIV command message if AMR is used *and* it is
desired to have working TFO-AMR per TS 28.062.  3GPP Rel4 added
MultiRate Control (not to be confused with MultiRate Configuration)
and Supported Codec Types IEs; Rel5 added TFO transparent container.
It is now known that Nokia CX BTS (MetroSite) don't support SCT IE -
let us now try MultiRate Control plus TFO transparent container,
which seems to be the Rel5 way of doing TFO-AMR.

Add vty options that enable sending of each of these IEs individually,
to facilitate experiments with different combinations on real E1 BTS
hardware.

Change-Id: Ib66bdf5f15717eb3f74fee34e0f436fac2241356