vty: rm misleading warnings about DTX on E1 BTS
Since the days of pre-split OpenBSC, whenever an operator would
enable DTX (either UL or DL) in BSC vty on an E1 BTS, there was
a warning message emitted, saying that this configuration is
"neither supported nor tested".  While those statements may have
been correct in 2016 when that code was added (commit 6079528b),
they are no longer correct in the present: both DTXu and DTXd
work correctly now on E1 BTS that follow TS 48.060 or 48.061,
with all 4 supported speech codecs: FR, HR, EFR and AMR.

Full-featured, production-quality operation with E1 BTS requires
use of tw-e1abis-mgw in the place of osmo-mgw - but even when
legacy OsmoMGW-E1 is used, it is misleading to say that DTX is
"neither supported nor tested", as the architectural shortcomings
of OsmoMGW-E1 (lack of TFO transform for non-AMR codecs and lack
of any support for AMR) are not specific to DTX per se.

Change-Id: I63626b554d38b1ab094838374968e64193da1c11